+++ src/components/SpyPlayer.tsx
@@ -1,15 +1,19 @@
 import { useSelector } from "react-redux";
 import { RootState } from "../store/store";
 import { motion } from "framer-motion";
-export default function SpySpyComponent() {
-  const spyIndex = useSelector(
-    (state: RootState) => state.PlayerSlice.spyIndex as number
+export default function SpyPlayer() {
+  const currentPlayerIndex = useSelector(
+    (state: RootState) => state.PlayerSlice.currentPlayerIndex
+  );
+  const location = useSelector(
+    (state: RootState) => state.PlayerSlice.location
   );
   const animationVariants = {
     initial: { opacity: 0, y: 20 },
     animate: { opacity: 1, y: 0 },
     exit: { opacity: 0, y: -20 },
   };
+
   return (
     <motion.div
       variants={animationVariants}
@@ -17,14 +21,10 @@ export default function SpySpyComponent() {
       animate="animate"
       exit="exit"
     >
-      <motion.p
-        className="text-red-300 font-bold"
-        animate={{ scale: 1.1 }}
-        transition={{ delay: 0.5 }}
-        data-testid="spyAnnouceLabel"
-      >
-        Người chơi số {spyIndex + 1}, bạn là điệp viên{"🤫"}
-      </motion.p>
+      <p className="font-bold" data-testid="playerAnnouceLabel">
+        Người chơi số {currentPlayerIndex + 1}
+      </p>
+      <p>Địa điểm là: {location}</p>
     </motion.div>
   );
 }
