+++ src/components/SpyReady.tsx
@@ -1,19 +1,16 @@
 import { useSelector } from "react-redux";
 import { RootState } from "../store/store";
 import { motion } from "framer-motion";
-export default function SpyPlayerComponent() {
+
+export default function SpyReady() {
   const currentPlayerIndex = useSelector(
     (state: RootState) => state.PlayerSlice.currentPlayerIndex
-  );
-  const location = useSelector(
-    (state: RootState) => state.PlayerSlice.location
   );
   const animationVariants = {
     initial: { opacity: 0, y: 20 },
     animate: { opacity: 1, y: 0 },
     exit: { opacity: 0, y: -20 },
   };
-
   return (
     <motion.div
       variants={animationVariants}
@@ -21,10 +18,7 @@ export default function SpyPlayerComponent() {
       animate="animate"
       exit="exit"
     >
-      <p className="font-bold" data-testid="playerAnnouceLabel">
-        Người chơi số {currentPlayerIndex + 1}
-      </p>
-      <p>Địa điểm là: {location}</p>
+      <p>Người chơi số: {currentPlayerIndex + 1} sẵn sàng chưa?</p>
     </motion.div>
   );
 }
