+++ src/features/PlayerSlice.spec.ts
@@ -27,17 +27,19 @@ describe("PlayerSlice's Reducer", () => {
   });
   describe("assignSpy reducer", () => {
     it("Should assign a spy", () => {
+      vi.spyOn(global.Math, "random").mockReturnValueOnce(0.1);
       const state = { ...initialState, value: 4 };
       const newState = playerReducer(state, assignSpy(4));
-      expect(newState.spyIndex).toBeGreaterThanOrEqual(0);
-      expect(newState.spyIndex).toBeLessThan(4);
+      expect(newState.spyIndex).toBe(0);
     });
   });
   describe("assignLocation reducer", () => {
     it("Should assign a random location in list", () => {
+      vi.spyOn(global.Math, "random").mockReturnValueOnce(0.01);
       const state = { ...initialState };
       const newState = playerReducer(state, assignLocation());
-      expect(locationList).toContain(newState.location);
+      expect(newState.location).toBe(locationList[0])
+
     });
   });
   describe("nextPlayer reducer", () => {
