+++ src/stories/SpyReady.stories.tsx
@@ -1,12 +1,12 @@
 import { Provider } from "react-redux";
 import { store } from "../store/store";
 import { Meta, StoryObj } from "@storybook/react";
-import SpySpyComponent from "../components/SpySpyComponent";
 import SpyAnnounceRole from "../components/SpyAnnounceRole";
+import SpyReady from "../components/SpyReady";
 
 const meta: Meta<typeof SpyAnnounceRole> = {
   title: "Components/Announce Player",
-  component: SpySpyComponent,
+  component: SpyReady,
   parameters: { layout: "centered" },
   decorators: [
     (Story) => (
@@ -18,4 +18,4 @@ const meta: Meta<typeof SpyAnnounceRole> = {
 };
 export default meta;
 type Story = StoryObj<typeof meta>;
-export const SpyAnnounce: Story = {};
+export const ReadyAnnounce: Story = {};
